+++ acceleratewp-debugger.php
@@ -3,7 +3,7 @@
  * Plugin Name:       AccelerateWP Debugger
  * Plugin URI:        https://wpeasy.au/
  * Description:       Svelte-based admin interface for managing AccelerateWP Preloader.
- * Version:           1.0.1
+ * Version:           1.0.2
  * Requires at least: 6.5
  * Requires PHP:      8.0
  * Author:            WPEasy
@@ -21,7 +21,7 @@ namespace AB\AccelerateWPDebugger;
 
 defined('ABSPATH') || exit;
 
-define('AWPD_VERSION', '1.0.1');
+define('AWPD_VERSION', '1.0.2');
 define('AWPD_PLUGIN_FILE', __FILE__);
 define('AWPD_PLUGIN_DIR', plugin_dir_path(__FILE__));
 define('AWPD_PLUGIN_URL', plugin_dir_url(__FILE__));